Move theme code and clean up
