+++ libs/anaconda-assistant-conda/src/anaconda_assistant_conda/rich_customizations/test_md.py
@@ -1,14 +1,11 @@
 import rich
-from rich.markdown import Markdown
 from rich.console import Console
 from rich.color import Color
 
-from .custom_theme import custom_theme
-from .no_padding_code_block import override_markdown_formatting
+from .console_theme import console_theme
+from .md import MyMarkdown
 
-override_markdown_formatting(Markdown)
-
-console = Console(theme=custom_theme)
+console = Console(theme=console_theme)
 
 markdown = """\
 Deserunt labore do elit mollit ut. Dolore Lorem quis non esse id nisi proident id esse tempor Lorem amet exercitation. Qui laborum ut mollit laborum proident. Qui elit mollit cupidatat ipsum. Voluptate aliquip elit consectetur dolor voluptate. Dolor labore est consectetur ipsum officia adipisicing eiusmod eiusmod velit dolor sunt sit labore. Qui cupidatat ea consequat adipisicing Lorem nostrud dolore. Aliqua dolore non magna Lorem elit nulla quis ullamco mollit ad mollit labore eiusmod. Non qui amet ipsum aliqua consequat nisi veniam officia. Id quis labore minim consectetur consectetur enim culpa quis cupidatat cupidatat voluptate aute non. Esse aute fugiat do eu excepteur sint dolore Lorem dolor.
@@ -38,4 +35,4 @@ This is a line with inline code `conda install pandas`
 Deserunt labore do elit mollit ut. Dolore Lorem quis non esse id nisi proident id esse tempor Lorem amet exercitation. Qui laborum ut mollit laborum proident. Qui elit mollit cupidatat ipsum. Voluptate aliquip elit consectetur dolor voluptate. Dolor labore est consectetur ipsum officia adipisicing eiusmod eiusmod velit dolor sunt sit labore. Qui cupidatat ea consequat adipisicing Lorem nostrud dolore. Aliqua dolore non magna Lorem elit nulla quis ullamco mollit ad mollit labore eiusmod. Non qui amet ipsum aliqua consequat nisi veniam officia. Id quis labore minim consectetur consectetur enim culpa quis cupidatat cupidatat voluptate aute non. Esse aute fugiat do eu excepteur sint dolore Lorem dolor.
 """
 
-console.print(Markdown(markdown))
+console.print(MyMarkdown(markdown))
